Seperate entity creation concerns from notestore
I'm not sure the Api namespace is a good descriptor of what
it contains but I like this design much better. The api
folder/namespace can be renamed but I feel that we should
pull much of the data transformation responsibility out
of the NoteStore and keep it a thin wrapper around
the thrift calls. Tests are still passing.